# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/iam/identity": {
      "post": {
        "description": "Creates a new user in the system with the provided information including contact details, roles, and permissions.",
        "operationId": "IdentityController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User has been successfully created."
          },
          "400": {
            "description": "Bad Request - Invalid input data."
          },
          "409": {
            "description": "Conflict - User with this email already exists."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a new administrator user",
        "tags": [
          "Identity Management"
        ]
      }
    },
    "/iam/identity/admin": {
      "get": {
        "operationId": "IdentityController_listAdministrators",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Administrators have been successfully retrieved."
          },
          "400": {
            "description": "Bad Request - Invalid input data."
          },
          "401": {
            "description": "Unauthorized - Invalid credentials."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "List administrators",
        "tags": [
          "Identity Management"
        ]
      }
    },
    "/iam/identity/{id}": {
      "get": {
        "description": "Retrieves a user by their unique identifier.",
        "operationId": "IdentityController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users have been successfully retrieved."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Get a user by id",
        "tags": [
          "Identity Management"
        ]
      },
      "patch": {
        "description": "Updates a user by their unique identifier.",
        "operationId": "IdentityController_updateUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been successfully updated."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Update a user by id",
        "tags": [
          "Identity Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Support Service API",
    "description": "The Support Service API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "EmailDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com",
            "format": "email"
          }
        },
        "required": [
          "address"
        ]
      },
      "PhoneDto": {
        "type": "object",
        "properties": {
          "dialCode": {
            "type": "string",
            "description": "Country dial code",
            "example": "+1",
            "pattern": "^\\+[1-9]\\d{0,3}$"
          },
          "number": {
            "type": "string",
            "description": "Phone number without country code",
            "example": "5551234567",
            "pattern": "^[0-9]{7,15}$"
          }
        },
        "required": [
          "dialCode",
          "number"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of the user",
            "example": "John",
            "minLength": 1,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user",
            "example": "Doe",
            "minLength": 1,
            "maxLength": 50
          },
          "username": {
            "type": "string",
            "description": "Username of the user",
            "example": "john.doe",
            "minLength": 1,
            "maxLength": 50
          },
          "email": {
            "description": "Email address of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/EmailDto"
              }
            ]
          },
          "phone": {
            "description": "Phone information of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhoneDto"
              }
            ]
          },
          "roles": {
            "description": "List of roles assigned to the user",
            "example": [
              "admin",
              "user-manager",
              "support"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "teams": {
            "description": "List of teams the user belongs to",
            "example": [
              "support",
              "development",
              "qa"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "includedPermissions": {
            "description": "List of additional permissions specifically granted to the user",
            "example": [
              "user.create",
              "user.read",
              "user.update",
              "report.generate"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "excludedPermissions": {
            "description": "List of permissions specifically denied to the user (overrides role permissions)",
            "example": [
              "user.delete",
              "system.admin"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "includedScopes": {
            "description": "List of access scopes granted to the user",
            "example": [
              "admin",
              "user-management",
              "reporting"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "excludedScopes": {
            "description": "List of access scopes denied to the user",
            "example": [
              "billing",
              "system-configuration"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "username",
          "email",
          "phone"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "The first name of the user",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the user",
            "example": "Doe"
          },
          "email": {
            "description": "Email information of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/EmailDto"
              }
            ]
          },
          "phone": {
            "description": "Phone information of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/PhoneDto"
              }
            ]
          },
          "roles": {
            "description": "The roles of the user",
            "example": [
              "admin",
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "teams": {
            "description": "The teams of the user",
            "example": [
              "team1",
              "team2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "includedPermissions": {
            "description": "The included permissions of the user",
            "example": [
              "permission1",
              "permission2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludedPermissions": {
            "description": "The excluded permissions of the user",
            "example": [
              "permission3",
              "permission4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "includedScopes": {
            "description": "The included scopes of the user",
            "example": [
              "scope1",
              "scope2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludedScopes": {
            "description": "The excluded scopes of the user",
            "example": [
              "scope3",
              "scope4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "firstName",
          "lastName",
          "roles",
          "teams",
          "includedPermissions"
        ]
      },
...
    }
  }
}